requirements.txt
opencv-python~=4.12.0.88
pytesseract~=0.3.13
PyMuPDF~=1.26.5
numpy~=2.2.6
pillow~=12.0.0
tqdm~=4.67.1